refactor(transport): extract Zenoh connection-with-timeout logic
Moves connectWithTimeout and the ConnectionResult helper into a focused
extension file. Drops the now-unnecessary 'private' on the stored client
property to internal (default), which is invisible from the public API.
